move common sca code to the common folder
