+++ checkov/common/bridgecrew/vulnerability_scanning/integrations/docker_image_scanning.py
@@ -3,11 +3,13 @@ from __future__ import annotations
 import logging
 import os
 from pathlib import Path
-from typing import Any
+from typing import Any, TYPE_CHECKING
 
-from checkov.common.bridgecrew.platform_integration import BcPlatformIntegration
 from checkov.common.bridgecrew.vulnerability_scanning.integrations.twistcli import TwistcliIntegration
 from checkov.common.util.str_utils import removeprefix
+
+if TYPE_CHECKING:
+    from checkov.common.bridgecrew.platform_integration import BcPlatformIntegration
 
 
 class DockerImageScanningIntegration(TwistcliIntegration):
