+++ checkov/common/bridgecrew/vulnerability_scanning/integrations/twistcli.py
@@ -5,12 +5,15 @@ import platform
 import stat
 from abc import abstractmethod, ABC
 from pathlib import Path
-from typing import Dict, Any, List
+from typing import Dict, Any, List, TYPE_CHECKING
 from datetime import datetime, timedelta
 
-from checkov.common.bridgecrew.platform_integration import bc_integration, BcPlatformIntegration
+from checkov.common.bridgecrew.platform_integration import bc_integration
 from checkov.common.util.data_structures_utils import merge_dicts
 from checkov.common.util.http_utils import get_default_post_headers, request_wrapper, aiohttp_client_session_wrapper
+
+if TYPE_CHECKING:
+    from checkov.common.bridgecrew.platform_integration import BcPlatformIntegration
 
 
 class TwistcliIntegration(ABC):
