+++ checkov/common/sca/commons.py
@@ -1,6 +1,8 @@
 from __future__ import annotations
 from checkov.common.output.common import ImageDetails
 
+UNFIXABLE_VERSION = "N/A"
+
 
 def get_file_path_for_record(rootless_file_path: str) -> str:
     return f"/{rootless_file_path}"
