+++ checkov/sca_image/runner.py
@@ -20,6 +20,7 @@ from checkov.common.output.report import Report, merge_reports
 from checkov.common.bridgecrew.check_type import CheckType
 from checkov.common.output.common import ImageDetails
 from checkov.common.runners.base_runner import filter_ignored_paths, strtobool
+from checkov.common.sca.output import parse_vulns_to_records
 from checkov.common.util.file_utils import compress_file_gzip_base64
 from checkov.common.util.dockerfile import is_docker_file
 from checkov.common.typing import _LicenseStatus
@@ -213,15 +214,17 @@ class Runner(PackageRunner):
                     pass
             rootless_file_path = dockerfile_path.replace(Path(dockerfile_path).anchor, "", 1)
 
-            self.parse_vulns_to_records(
+            parse_vulns_to_records(
                 report=report,
+                check_class=self._check_class,
                 scanned_file_path=os.path.abspath(dockerfile_path),
                 rootless_file_path=f"{rootless_file_path} ({image.name} lines:{image.start_line}-{image.end_line} ({image_id}))",
                 runner_filter=runner_filter,
                 vulnerabilities=vulnerabilities,
                 packages=[],
                 license_statuses=[],
-                image_details=image_details
+                image_details=image_details,
+                report_type=self.report_type,
             )
 
             return report
@@ -235,14 +238,16 @@ class Runner(PackageRunner):
             self.raw_report = scan_result
             result = scan_result.get('results', [{}])[0]
             vulnerabilities = result.get("vulnerabilities") or []
-            self.parse_vulns_to_records(
+            parse_vulns_to_records(
                 report=report,
+                check_class=self._check_class,
                 scanned_file_path=os.path.abspath(dockerfile_path),
                 rootless_file_path=f"{dockerfile_path} ({image.name} lines:{image.start_line}-{image.end_line} ({image_id}))",
                 runner_filter=runner_filter,
                 vulnerabilities=vulnerabilities,
                 packages=[],
                 license_statuses=[],
+                report_type=self.report_type,
             )
         else:
             logging.info(f"No cache hit for image {image.name}")
@@ -297,15 +302,17 @@ class Runner(PackageRunner):
                 # Path.is_relative_to() was implemented in Python 3.9
                 pass
         rootless_file_path = dockerfile_path.replace(Path(dockerfile_path).anchor, "", 1)
-        self.parse_vulns_to_records(
+        parse_vulns_to_records(
             report=report,
+            check_class=self._check_class,
             scanned_file_path=os.path.abspath(dockerfile_path),
             rootless_file_path=f"{rootless_file_path} ({image_id})",
             runner_filter=runner_filter,
             vulnerabilities=vulnerabilities,
             packages=packages,
             license_statuses=license_statuses,
-            image_details=image_details
+            image_details=image_details,
+            report_type=self.report_type,
         )
         return report
